[S07P22D203-581|BACK] 테스트
Former-commit-id: aabd95b45ce22636cf3adf3b857da5a7fc53fd12 [formerly 3d521cc3b7dab9f2914bde7ed62f2e076fba6254]
Former-commit-id: ee68709d442478ac1dbdfbea6c5f8c9d26bb5291

+++ BACK/src/main/java/com/ihagong/momssok/service/PromiseService.java
@@ -23,10 +23,15 @@ public class PromiseService {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         Map<Boolean, Object> result = new HashMap<>();
         Map<String, Object> resultBody = new HashMap<>();
-
+        if(promise.getPromiseItems()==null){
+            resultBody.put("message", "promiseItems 입력 오류");
+            result.put(false, resultBody);
+            return result;
+        }
         if(promiseMapper.selectPromise(email+"_"+promise.getName())!=null){
             promiseMapper.deletePromise(email+"_"+promise.getName());
         }
+
         int index=1;
         for(PromiseItemDto item:promise.getPromiseItems()){
             item.setDone(0);
